Move functions from  to .
